+++ vault_cli/client.py
@@ -5,6 +5,7 @@ import pathlib
 from typing import Iterable, Optional, Tuple, Type
 
 import hvac
+import jinja2
 import requests.packages.urllib3
 
 from vault_cli import exceptions, sessions, settings, types, utils
@@ -88,6 +89,7 @@ class VaultClientBase:
         username: Optional[str] = settings.DEFAULTS.username,
         password: Optional[str] = settings.DEFAULTS.password,
         safe_write: bool = settings.DEFAULTS.safe_write,
+        render: bool = settings.DEFAULTS.render,
     ):
         """
         All parameters are mandatory but may be None
@@ -102,6 +104,7 @@ class VaultClientBase:
         self.username = username
         self.password = password
         self.safe_write = safe_write
+        self.render = render
 
     def auth(self):
         verify_ca_bundle = self.verify
@@ -154,7 +157,9 @@ class VaultClientBase:
         """
         pass
 
-    def _browse_recursive_secrets(self, path: str) -> Iterable[str]:
+    def _browse_recursive_secrets(
+        self, path: str, render: bool = True
+    ) -> Iterable[str]:
         """
         Given a secret or folder path, return the path of all secrets
         under it (or the path itself)
@@ -181,10 +186,10 @@ class VaultClientBase:
                 yield key_url
                 continue
 
-            for sub_path in self._browse_recursive_secrets(key_url):
+            for sub_path in self._browse_recursive_secrets(key_url, render=render):
                 yield sub_path
 
-    def get_all_secrets(self, *paths: str) -> types.JSONDict:
+    def get_all_secrets(self, *paths: str, render: bool = True) -> types.JSONDict:
         """
         Takes several paths, return the nested dict of all secrets below
         those paths
@@ -193,35 +198,45 @@ class VaultClientBase:
         result: types.JSONDict = {}
 
         for path in paths:
-            path_dict = self.get_secrets(path)
+            path_dict = self.get_secrets(path, render=render)
 
             result.update(utils.path_to_nested(path_dict))
 
         return result
 
-    def get_secrets(self, path: str) -> types.JSONDict:
+    def get_secrets(self, path: str, render: bool = True) -> types.JSONDict:
         """
         Takes a single path an return a path dict with all the secrets
         below this path, recursively
         """
-        secrets_paths = self._browse_recursive_secrets(path=path)
-        return {subpath: self.get_secret(path=subpath) for subpath in secrets_paths}
+        secrets_paths = self._browse_recursive_secrets(path=path, render=render)
+        result: types.JSONDict = {}
+        for subpath in secrets_paths:
+            try:
+                result[subpath] = self.get_secret(path=subpath, render=render)
+            except exceptions.VaultAPIException:
+                result[subpath] = "<error while retrieving secret>"
 
-    def get_all(self, *args, **kwargs):
-        """
-        Synonym to get_all_secrets. Can be removed on 0.6.0.
-        """
-        logger.warning(
-            "Using deprecated 'get_all' method. Use 'get_all_secrets' instead."
-        )
-        return self.get_all_secrets(*args, **kwargs)
+        return result
+
+    def list_secrets(self, path: str) -> Iterable[str]:
+        return self._list_secrets(path=path)
+
+    def get_secret(self, path: str, render: bool = True) -> types.JSONValue:
+        secret = self._get_secret(path=path)
+        if render and self.render:
+            secret = self._render_template_value(secret)
+        return secret
+
+    def delete_secret(self, path: str) -> None:
+        return self._delete_secret(path=path)
 
     def delete_all_secrets_iter(self, *paths: str) -> Iterable[str]:
         """
         Recursively deletes all the secrets at the given paths.
         """
         for path in paths:
-            secrets_paths = self._browse_recursive_secrets(path=path)
+            secrets_paths = self._browse_recursive_secrets(path=path, render=False)
             for secret_path in secrets_paths:
                 yield secret_path
                 self.delete_secret(secret_path)
@@ -236,7 +251,7 @@ class VaultClientBase:
         self, source: str, dest: str, force: Optional[bool] = None
     ) -> Iterable[Tuple[str, str]]:
 
-        source_secrets = self.get_secrets(path=source)
+        source_secrets = self.get_secrets(path=source, render=False)
 
         for old_path, secret in source_secrets.items():
             new_path = dest + old_path[len(source) :]
@@ -248,21 +263,44 @@ class VaultClientBase:
             self.delete_secret(old_path)
 
     def move_secrets(
-        self, source: str, dest: str, force: bool = False, generator: bool = False
+        self,
+        source: str,
+        dest: str,
+        force: Optional[bool] = None,
+        generator: bool = False,
     ) -> Iterable[Tuple[str, str]]:
         iterator = self.move_secrets_iter(source=source, dest=dest, force=force)
         if generator:
             return iterator
         return list(iterator)
 
+    template_prefix = "!template!"
+
+    def _render_template_value(self, secret: types.JSONValue) -> types.JSONValue:
+        if not isinstance(secret, str):
+            return secret
+
+        if not secret.startswith(self.template_prefix):
+            return secret
+
+        return self.render_template(secret[len(self.template_prefix) :], render=False)
+
+    def render_template(self, template: str, render: bool = True) -> str:
+        def vault(path):
+            try:
+                return self.get_secret(path, render=render)
+            except exceptions.VaultException:
+                raise exceptions.VaultRenderTemplateError(f"'{path}' not found")
+
+        return jinja2.Template(template).render(vault=vault)
+
     def set_secret(
         self, path: str, value: types.JSONValue, force: Optional[bool] = None
     ) -> None:
-
         force = self.get_force(force)
 
         try:
-            existing_value = self.get_secret(path=path)
+            existing_value = self.get_secret(path=path, render=False)
         except exceptions.VaultSecretNotFound:
             pass
         except exceptions.VaultForbidden:
@@ -289,7 +327,7 @@ class VaultClientBase:
         path = path.rstrip("/")
         for parent in list(pathlib.PurePath(path).parents)[:-1]:
             try:
-                self.get_secret(str(parent))
+                self.get_secret(str(parent), render=False)
             except exceptions.VaultSecretNotFound:
                 pass
             except exceptions.VaultForbidden:
@@ -321,19 +359,22 @@ class VaultClientBase:
     def _authenticate_userpass(self, username: str, password: str) -> None:
         raise NotImplementedError
 
-    def list_secrets(self, path: str) -> Iterable[str]:
+    def _list_secrets(self, path: str) -> Iterable[str]:
         raise NotImplementedError
 
-    def get_secret(self, path: str) -> types.JSONValue:
+    def _get_secret(self, path: str) -> types.JSONValue:
         raise NotImplementedError
 
-    def delete_secret(self, path: str) -> None:
+    def _delete_secret(self, path: str) -> None:
         raise NotImplementedError
 
     def _set_secret(self, path: str, value: types.JSONValue) -> None:
         raise NotImplementedError
 
     def lookup_token(self) -> types.JSONDict:
+        return self._lookup_token()
+
+    def _lookup_token(self) -> types.JSONDict:
         raise NotImplementedError
 
 
@@ -389,21 +430,21 @@ class VaultClient(VaultClientBase):
         self.client.auth_tls()
 
     @handle_errors()
-    def list_secrets(self, path: str) -> Iterable[str]:
+    def _list_secrets(self, path: str) -> Iterable[str]:
         secrets = self.client.list(self.base_path + path)
         if not secrets:
             return []
         return secrets["data"]["keys"]
 
     @handle_errors()
-    def get_secret(self, path: str) -> types.JSONValue:
+    def _get_secret(self, path: str) -> types.JSONValue:
         secret = self.client.read(self.base_path + path)
         if not secret:
             raise exceptions.VaultSecretNotFound()
         return secret["data"]["value"]
 
     @handle_errors()
-    def delete_secret(self, path: str) -> None:
+    def _delete_secret(self, path: str) -> None:
         self.client.delete(self.base_path + path)
 
     @handle_errors()
@@ -411,7 +452,7 @@ class VaultClient(VaultClientBase):
         self.client.write(self.base_path + path, value=value)
 
     @handle_errors()
-    def lookup_token(self) -> types.JSONDict:
+    def _lookup_token(self) -> types.JSONDict:
         return self.client.lookup_token()
 
     def __exit__(self, exc_type, exc_value, traceback):