+++ CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ------------------
 
 - Add `vault lookup-token` (#77)
+- Remove `client.get_all()` (after deprecation period)
+- Add template values and `--render` features in client and cli (#65)
 
 0.6.0 (2019-15-06)
 ------------------